[probability] Switch ParameterController to double
Even though the distributions store their parameters as double, they
used to come from a FloatParameterController<float>, which led to
imprecisions.

+++ apps/probability/distribution/chi_squared_distribution.h
@@ -11,7 +11,7 @@ public:
   static constexpr int k_maxRegularizedGammaIterations = 1000;
   static constexpr double k_regularizedGammaPrecision = DBL_EPSILON;
 
-  ChiSquaredDistribution() : OneParameterDistribution(1.0f) {}
+  ChiSquaredDistribution() : OneParameterDistribution(1.0) {}
   I18n::Message title() override { return I18n::Message::ChiSquaredDistribution; }
   Type type() const override { return Type::ChiSquared; }
   bool isContinuous() const override { return true; }
@@ -27,7 +27,7 @@ public:
     return I18n::Message::DegreesOfFreedomDefinition;
   }
   float evaluateAtAbscissa(float x) const override;
-  bool authorizedValueAtIndex(float x, int index) const override;
+  bool authorizedValueAtIndex(double x, int index) const override;
   double cumulativeDistributiveFunctionAtAbscissa(double x) const override;
   double cumulativeDistributiveInverseForProbability(double * probability) override;
 private:
